# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EVA Auth API",
        "contact": {
            "email": "gurkalov.dev@gmail.com"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://auth.evarun.ru"
        }
    ],
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "description": "Login as user.",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "Creds for authorization.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginCreds"
                            },
                            "example": {
                                "email": "test@email.com",
                                "password": "secret"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authorize user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserApiKey"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "description": "Register new user.",
                "operationId": "registerUser",
                "requestBody": {
                    "description": "Creds for registeration.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterCreds"
                            },
                            "example": {
                                "email": "new-user-test@email.com",
                                "password": "secret",
                                "name": "Tim Cook"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserApiKey"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "description": "Returns a user info which auth for token",
                "operationId": "profileUser",
                "responses": {
                    "200": {
                        "description": "User response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "description": "Update user profile.",
                "operationId": "updateProfile",
                "requestBody": {
                    "description": "Creds for registeration.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            },
                            "example": {
                                "email": "api-test@email.com",
                                "password": "secret",
                                "name": "Api Tim Cook",
                                "status": "free"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Returns all users",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "maximum number of results to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "example": "10"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "User bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Returns a user based on a single ID",
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ErrorModel": {
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "LoginCreds": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "string",
                        "example": "test@email.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "string",
                        "example": "secret"
                    }
                },
                "type": "object"
            },
            "RegisterCreds": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "string",
                        "example": "test@email.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "string",
                        "example": "secret"
                    },
                    "name": {
                        "type": "string",
                        "format": "string",
                        "example": "Tom Sand"
                    }
                },
                "type": "object"
            },
            "UserApiKey": {
                "required": [
                    "id",
                    "api_key"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "api_key": {
                        "type": "string",
                        "format": "string",
                        "example": "em1JbEVqSnZlR0lPMlozenZ5YmpPUWdKSURiVGpnZmg="
                    }
                },
                "type": "object"
            },
            "User": {
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "admin": {
                        "type": "boolean",
                        "format": "boolean",
                        "example": false
                    },
                    "name": {
                        "type": "string",
                        "format": "string",
                        "example": "User Name"
                    },
                    "status": {
                        "type": "string",
                        "format": "string",
                        "example": "thebest"
                    },
                    "options": {
                        "type": "object",
                        "format": "array",
                        "items": {
                            "type": "array",
                            "items": {},
                            "example": "{}"
                        },
                        "example": "{}"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "string",
                        "example": "2019-01-26 20:00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "string",
                        "example": "2019-01-26 20:00:57"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}